projects/MLOps-orchestration/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
prefect>=3.0.0
mlflow>=3.0.0

# Google APIs
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Additional utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0


